Switch to using SetupDi for enumerating HID devices
Done to more easily get access to HW IDs through the SPDRP_HARDWAREID property. The HW IDs are returned on a "HID\VID_0B0E&PID_0312&REV_0213&MI_03&Col02" format.

WARNING: Instance information seem to be lacking in the HW IDs, so it's not possible to uniquely identify a given device based on the HW ID.